WIP: per-block filters, protobuf, address finder
As a Bloom filter's size just depends on the element count for a given false positive rate and the exact block also has to be determined somehow anyway, the filters are now per block
Each block is identified by its hash
The serialization uses protocol buffers
Some overhead is stored in form of the used hash function which never changes but that's 2 bytes per block only. Also the block hash could be stored without leading zeros.
WIP: it does currently not work! addresses are not found! Bug!